+++ src/components/Footer/FooterImage.jsx
